+++ x/erc20/keeper/keeper.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cosmos/evm/x/erc20/types"
 	transferkeeper "github.com/cosmos/evm/x/ibc/transfer/keeper"
 
+	"cosmossdk.io/core/address"
 	"cosmossdk.io/log"
 	storetypes "cosmossdk.io/store/types"
 
@@ -19,6 +20,7 @@ type Keeper struct {
 	cdc      codec.BinaryCodec
 	// the address capable of executing a MsgUpdateParams message. Typically, this should be the x/gov module account.
 	authority sdk.AccAddress
+	addrCodec address.Codec
 
 	accountKeeper  types.AccountKeeper
 	bankKeeper     types.BankKeeper
@@ -52,6 +54,7 @@ func NewKeeper(
 		evmKeeper:      evmKeeper,
 		stakingKeeper:  sk,
 		transferKeeper: transferKeeper,
+		addrCodec:      ak.AddressCodec(),
 	}
 }